+++ sauron-node-script.js
@@ -1,0 +1,7 @@
+'use strict';
+module.exports = {
+  extends: ['./sauron-node', 'plugin:node/recommended-script'],
+  rules: {
+    'import/no-commonjs': 'off'
+  }
+};